+++ tests/testthat/test-backend-sqlite.R
@@ -37,6 +37,11 @@ test_that("custom aggregates translated", {
 
   expect_equal(translate_sql(median(x, na.rm = TRUE), window = FALSE), sql('MEDIAN(`x`)'))
   expect_equal(translate_sql(sd(x, na.rm = TRUE), window = FALSE), sql('STDEV(`x`)'))
+
+  expect_snapshot({
+    (expect_error(translate_sql(quantile(x, 0.5, na.rm = TRUE), window = FALSE)))
+    (expect_error(translate_sql(quantile(x, 0.5, na.rm = TRUE), window = TRUE)))
+  })
 })
 
 test_that("custom SQL translation", {